requirements.txt
google-cloud-translate==0.21.0
nltk
google-cloud-language
